corp_agent_backend/requirements.txt
# Web Framework
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
alembic>=1.10.0

# Authentication & Security
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
fastapi-limiter>=0.1.5
redis>=4.5.0

# LLM & AI Components
langchain>=0.0.267
langchain-huggingface>=0.0.1
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0
torch>=2.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
huggingface-hub>=0.16.0
bitsandbytes>=0.41.0
einops>=0.6.0

# Monitoring & Logging
prometheus-fastapi-instrumentator>=0.21.0
python-dotenv>=1.0.0

# Testing
pytest>=7.3.0
httpx>=0.24.0
